+++ package.json
@@ -7,11 +7,11 @@
     "@emotion/styled": "^10.0.23",
     "axios": "^0.19.0",
     "emotion-reset": "^2.0.3",
-    "firebase": "^7.4.0",
+    "firebase": "^7.5.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-emotion": "^10.0.0",
-    "react-hook-form": "^3.28.2",
+    "react-hook-form": "^3.28.11",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0"
