--- src/components/Items/Item.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-const Item = props => {
-  return (
-    <>
-      <div>Image of item</div>
-      <div>Item description</div>
-      <div>Item price</div>
-      <div>Item location?</div>
-    </>
-  );
-};
-
-export default Item;
