--- src/components/Items/Items.js
@@ -1,17 +1,0 @@
-import React from "react";
-
-import Item from "./Item";
-
-const Items = props => {
-  return (
-    <>
-      <div>List of items</div>
-      <div>Map through Items</div>
-      <div>
-        <Item />
-      </div>
-    </>
-  );
-};
-
-export default Items;
