+++ src/components/Layout/Navbar/SignedInLinks.js
@@ -1,6 +1,5 @@
 import React, { useContext } from "react";
 import { NavLink } from "react-router-dom";
-import styled from "@emotion/styled";
 
 import firebase from "../../../firebase/firebase";
 import "firebase/auth";
@@ -8,24 +7,6 @@ import userContext from "../../../context/user/UserContext";
 
 const SignedInLinks = () => {
   const Users = useContext(userContext);
-
-  const Ul = styled.ul`
-    list-style-type: none;
-    margin: 20px 0 0;
-    padding: 0;
-  `;
-
-  const Li = styled.li`
-    background-color: #ccc;
-    padding: 10px;
-    margin: 8px;
-    border-radius: 4px;
-    display: inline;
-  `;
-
-  const StyledNavLink = styled(NavLink)`
-    text-decoration: none;
-  `;
 
   const SignOut = () => {
     firebase.auth().signOut().then(function() {
@@ -39,25 +20,25 @@ const SignedInLinks = () => {
 
   return (
     <div>
-      <Ul>
+      <ul>
         {Users.state.user.rankrole === "farmer" || "admin" ? (
-          <Li>
-            <StyledNavLink to="/farmerdash">Farmer Dashboard</StyledNavLink>
-          </Li>
+          <li>
+            <NavLink to="/farmerdash">Farmer Dashboard</NavLink>
+          </li>
         ) : null}
-        <Li>
-          <StyledNavLink to="/">Locations</StyledNavLink>
-        </Li>
-        <Li>
-          <StyledNavLink to="/">Inventory</StyledNavLink>
-        </Li>
-        <Li>
-          <StyledNavLink to="/userdash">{Users.state.user.firstName}</StyledNavLink>
-        </Li>
-        <Li>
-          <StyledNavLink onClick={() => SignOut()} to="/">Logout</StyledNavLink>
-        </Li>
-      </Ul>
+        <li>
+          <NavLink to="/">Locations</NavLink>
+        </li>
+        <li>
+          <NavLink to="/">Inventory</NavLink>
+        </li>
+        <li>
+          <NavLink to="/userdash">{Users.state.user.firstName}</NavLink>
+        </li>
+        <li>
+          <NavLink onClick={() => SignOut()} to="/">Logout</NavLink>
+        </li>
+      </ul>
     </div>
   );
 };
