+++ src/components/Layout/Navbar/SignedOutLinks.js
@@ -1,32 +1,12 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import styled from "@emotion/styled";
 
 import GoogleSigninFunc from "../../../firebase/Google";
 
-
-const SignedOutLinks = () => {
-  const Ul = styled.ul`
-    list-style-type: none;
-    margin: 20px 0 0;
-    padding: 0;
-  `;
-
-  const Li = styled.li`
-    background-color: #ccc;
-    padding: 10px;
-    margin: 8px;
-    border-radius: 4px;
-    display: inline;
-  `;
-
-  const StyledNavLink = styled(NavLink)`
-    text-decoration: none;
-  `;
-
+export default function SignedOutLinks() {
   return (
     <div>
-      <Ul>
+      <ul>
         {/* <Li>
           <StyledNavLink to="/">
             Register
@@ -37,12 +17,11 @@ const SignedOutLinks = () => {
             Sign In
           </StyledNavLink>
         </Li> */}
-        <Li>
+        <li>
           <GoogleSigninFunc />
-        </Li>
-      </Ul>
+        </li>
+      </ul>
     </div>
-  );
-};
+  )
+}
 
-export default SignedOutLinks;
